% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_handicaps.R
\name{calculate_handicaps}
\alias{calculate_handicaps}
\title{Use 5k time to predict running time using the Reigel Formula}
\usage{
calculate_handicaps(
  slowest_runner = as.numeric(lubridate::ms("30:00")),
  race_distance = 2.145
)
}
\arguments{
\item{slowest_runner}{the time of the slowest runner}

\item{race_distance}{the distance of the predicted race}
}
\value{

}
\description{
Use 5k time to predict running time using the Reigel Formula
}
